+++ tgbotapi.core/src/linuxX64Main/kotlin/bot/ktor/base/ActualPlatformClientCopy.kt
@@ -1,6 +1,8 @@
 package dev.inmo.tgbotapi.bot.ktor.base
 
+import dev.inmo.tgbotapi.utils.defaultKtorEngine
 import io.ktor.client.*
+import io.ktor.client.HttpClient
 import io.ktor.client.engine.curl.*
 
 /**
@@ -21,4 +23,8 @@ internal actual inline fun platformClientCopy(client: HttpClient): HttpClient = 
     HttpClient(Curl) {
         this.plusAssign(config)
     }
-} ?: throw IllegalArgumentException("On LinuxX64 TelegramBotAPI currently support only Curl Ktor HttpClient engine")
+} ?: HttpClient(
+        defaultKtorEngine
+) {
+    install(client)
+}